made contract card and modal same style as all the others
